13/main_test.go
package main

import (
	"fmt"
	"testing"
)

func TestStep1SampleInput(t *testing.T) {
	var tests = []struct {
		input    string
		expected int
	}{
		{`Button A: X+94, Y+34
Button B: X+22, Y+67
Prize: X=8400, Y=5400

Button A: X+26, Y+66
Button B: X+67, Y+21
Prize: X=12748, Y=12176

Button A: X+17, Y+86
Button B: X+84, Y+37
Prize: X=7870, Y=6450

Button A: X+69, Y+23
Button B: X+27, Y+71
Prize: X=18641, Y=10279`, 480},
		{`Button A: X+99, Y+0
Button B: X+1, Y+0
Prize: X=101, Y=0`, 5},
		{`Button A: X+19, Y+54
Button B: X+35, Y+17
Prize: X=17401, Y=6576`, 0},
		{`Button A: X+67, Y+11
Button B: X+16, Y+64
Prize: X=1755, Y=2067`, 0},
	}

	for i, tt := range tests {
		testname := fmt.Sprintf("%d", i)
		t.Run(testname, func(t *testing.T) {
			ans := runStep1(tt.input)
			if ans != tt.expected {
				t.Errorf(`TestStep1SampleInput got %d, expected %d`, ans, tt.expected)
			}
		})
	}
}
